third_party/WebKit/LayoutTests/media/video-controls.html

<!DOCTYPE html>
<title>Test "controls" attribute.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_true(video.controls);

    video.oncanplaythrough = t.step_func_done(function() {
        assert_true(video.controls);
        video.removeAttribute("controls");
        assert_false(video.controls);
        video.controls = true;
        assert_equals(video.getAttribute("controls"), "");
        video.controls = false;
        assert_false(video.controls);
        video.setAttribute("controls", "controls");
        assert_true(video.controls);
    });

    video.src = "content/test.ogv";
});
</script>
